Removed CFRelease call broken on OS X 10.12.
Addresses GitHub issue #3.

Since dmidecode is a short lived process, I'll wear
the risk of the dandgling handle, so that 10.12
users can at least make use of the binary.

Need to understand why the release call broke on 10.12.
Maybe something to do with privileged system calls, as
it works fine when running in the signed lldb debugger.